Revert "feature: 실패 처리 추가"
This reverts commit 494561254680b159f93155b34600b3cd8058923d.
