+++ src/main/java/com/example/pillyohae/global/distributedLock/DistributedLockAop.java
@@ -27,12 +27,6 @@ public class DistributedLockAop {
     private final RedissonClient redissonClient;
     private final AopForTransaction aopForTransaction;
 
-    /**
-     * AOP 적용 CLASS
-     * @param joinPoint 메소드 인자 변수명-변수값을 가져오고 어노테이션에 적용된 key 이름을 가져옴
-     * @return
-     * @throws Throwable
-     */
     @Around("@annotation(com.example.pillyohae.global.distributedLock.DistributedLock)")
     public Object lock(final ProceedingJoinPoint joinPoint) throws Throwable {
         MethodSignature signature = (MethodSignature) joinPoint.getSignature();
@@ -40,7 +34,7 @@ public class DistributedLockAop {
         DistributedLock distributedLock = method.getAnnotation(DistributedLock.class);
 
         String key = REDISSON_LOCK_PREFIX + CustomSpringELParser.getDynamicValue(signature.getParameterNames(), joinPoint.getArgs(), distributedLock.key());
-        RLock rLock = redissonClient.getLock(key);
+        RLock rLock = redissonClient.getLock(key);  // (1)
 
         try {
             boolean available = rLock.tryLock(distributedLock.waitTime(), distributedLock.leaseTime(), distributedLock.timeUnit());  // (2)
@@ -48,12 +42,12 @@ public class DistributedLockAop {
                 return false;
             }
 
-            return aopForTransaction.proceed(joinPoint);
+            return aopForTransaction.proceed(joinPoint);  // (3)
         } catch (InterruptedException e) {
             throw new InterruptedException();
         } finally {
             try {
-                rLock.unlock();
+                rLock.unlock();   // (4)
             } catch (IllegalMonitorStateException e) {
                 log.info("Redisson Lock Already UnLock {} {}",
                         kv("serviceName", method.getName()),
